% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{CreateModelCode}
\alias{CreateModelCode}
\alias{CreateModelCode.default}
\alias{ReadStanChunk}
\title{Create a Stan code file}
\usage{
CreateModelCode(dataStan, ...)

\method{CreateModelCode}{default}(dataStan, ...)

ReadStanChunk(prefix, type, suffix, isType = TRUE)
}
\arguments{
\item{dataStan}{a \code{list} output of a \code{\link[=CreateData]{CreateData()}} call.}

\item{...}{additional arguments to be passed to
\code{\link[cmdstanr:write_stan_file]{cmdstanr::write_stan_file()}}}
}
\value{
The path to a file with \code{stan} extension.
It contains the code with the specification of data structure, priors,
and log-likelihood of the given model.
}
\description{
\code{goldfish.latent} offers working version of the models to work with Stan.
Users should have installed \code{cmdstanr} package and \code{CmdStan} before.
Follow the instructions instructions from \code{cmdstanr} documentation.
}
\details{
\code{cmdstanr} functionalities makes possible for users to write a local copy
of the Stan code and modify it for other purposes.
The default in \code{\link[cmdstanr:write_stan_file]{cmdstanr::write_stan_file()}} is to write the model in a
temporal folder. Using the \code{dir} argument is possible to write the model
in a folder specifies by the user.
}
\examples{
\donttest{
library(goldfish)
library(cmdstanr)
data("Social_Evolution")
callNetwork <- defineNetwork(nodes = actors, directed = TRUE) |>
  linkEvents(changeEvent = calls, nodes = actors)
callsDependent <- defineDependentEvents(
  events = calls, nodes = actors, defaultNetwork = callNetwork
)
data2stan <- CreateDataModel(
  randomEffects = list(inertia ~ 1),
  fixedEffects = callsDependent ~ recip + trans
)

stanCode <- CreateModelCode(data2stan)
}
}
