% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups-methods.R
\name{gather_groups}
\alias{gather_groups}
\title{Gather data from multiple groups into a single object}
\usage{
gather_groups(groups_info, data, keep_x = NULL, scale = FALSE)
}
\arguments{
\item{groups_info}{A data frame with metadata about the groups. It must
contain at least a column \code{ixGroup} with the group identifier.}

\item{data}{A list of preprocessed data objects. Each element of the list
corresponds to a group and should be a \code{goldfish.latent.data} object
containing a \code{data_stan} list.}

\item{keep_x}{An optional character vector of column names to keep from
the \code{X_choice} matrix. If \code{NULL}, all columns are kept.}

\item{scale}{logical value. Whether to standardize the effect stats at
the end.}
}
\value{
A \code{goldfish.latent.data} object containing the aggregated data in the
\code{dataStan} list, along with metadata from the first group.
}
\description{
This function aggregates preprocessed data from a list of groups into a
single data structure suitable for a random effects by group model (using
Stan).
It combines matrices and vectors from each group, adjusting indices as
needed.
}
