% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups-methods.R
\name{gather_groups_to_json}
\alias{gather_groups_to_json}
\title{Gather data from multiple groups to a JSON file}
\usage{
gather_groups_to_json(
  groups_info,
  data,
  file_name,
  keep_x = NULL,
  scale = FALSE
)
}
\arguments{
\item{groups_info}{a data frame with the groups information.}

\item{data}{a list with the data from each group.}

\item{file_name}{the name of the file to save the data,
with \code{.json} extension.}

\item{keep_x}{a character vector with the names of the variables to keep.}

\item{scale}{a logical indicating whether to scale the data.}
}
\value{
a list with the data in long format.
}
\description{
Gather data from multiple groups to a JSON file
}
