% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{hmm_draws_to_label_switching}
\alias{hmm_draws_to_label_switching}
\title{Get Draws from the Posterior Distribution of a HMM-DyNAM}
\usage{
hmm_draws_to_label_switching(
  data_to_stan,
  cmdstan_samples,
  k_states = data_to_stan$dataStan$kS,
  type = c("2label.switching", "draws_matrix", "draws_array", "draws_df"),
  rescale = FALSE
)
}
\arguments{
\item{data_to_stan}{a object of class \code{"goldfish.latent.data"} outcome of
a call to \code{\link[=make_data_hmm]{make_data_hmm()}}.}

\item{cmdstan_samples}{a object of class \code{"CmdStanMCMC"} with the posterior
samples of a HMM from the \code{data_to_stan}.}

\item{k_states}{an integer with the number of states used to get samples
from the posterior distribution in \code{cmdstan_samples}.}

\item{type}{a character specifying weather the draws are shaped for the
\code{label.switching} package or a simple matrix format.}
}
\value{
a list with draws and additional information
}
\description{
Get Draws from the Posterior Distribution of a HMM-DyNAM
}
\examples{
\donttest{
library(goldfish)
data("Social_Evolution")
callNetwork <- make_network(nodes = actors, directed = TRUE) |>
  link_events(changeEvent = calls, nodes = actors)
callsDependent <- define_dependent_events(
  events = calls, nodes = actors, defaultNetwork = callNetwork
)
socialEvolutionData <- make_data(callsDependent)
data_to_stan <- make_data_hmm(
  rate_effects = callsDependent ~ indeg + outdeg,
  choice_effects = NULL,
  data = socialEvolutionData
)

drawsMCMC <- hmm_draws_to_label_switching(data_to_stan, cmdstanSamples)
}
}
