% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{hmm_post_processing}
\alias{hmm_post_processing}
\title{Post process MCMC samples from a HMM}
\usage{
hmm_post_processing(
  data_to_stan,
  cmdstan_samples,
  k_states = data_to_stan$data_stan$kS,
  type = c("both", "viterbi", "smoothProbs"),
  smooth_probs_states = c("joint", "marginal", "none"),
  cl = NULL,
  n_nodes = parallel::detectCores() - 1L
)
}
\arguments{
\item{data_to_stan}{a object of class \code{"goldfish.latent.data"} outcome of
a call to \code{\link[=make_data_hmm]{make_data_hmm()}}.}

\item{cmdstan_samples}{a object of class \code{"CmdStanMCMC"} with the posterior
samples of a HMM from the \code{data_to_stan}.}

\item{k_states}{an integer with the number of states used to get samples
from the posterior distribution in \code{cmdstan_samples}.}

\item{type}{a character specifying which algorithm to use for the post
processing of the posterior distribution samples.}

\item{smooth_probs_states}{a character specifying the way that state sample path
are drawn either using the smoothed probabilities \code{"marginal"} or
conditioning of the next sample state \code{"joint"}.
Only used when \code{type = "both"} or \code{type = "smoothProbs"}.}

\item{cl}{an optional \code{cluster} object to parallelize the
post-processing}

\item{n_nodes}{an integer specifying the number of nodes to use in the
parallelization.}
}
\value{
a list with the most probable path when Viterbi algorithm is used
and/or the smoothed probabilities and sample paths when the
Forward-Backward algorithm is used.
}
\description{
The post-processing allows to compute for each posterior draw:
the most probable path of the Hidden States using the Viterbi algorithm,
the smoothed probabilities of a Hidden State given the observed sequence
using the Forward-Backward algorithm, and a sample path using the smoothed
probabilities \code{smoothProbsSt = "marginal"} or taking into account the
conditional probability given the next state  \code{smoothProbsSt = "joint"} as
implemented in Stan \code{hmm_latent_rng()}.
}
\examples{
\donttest{
library(goldfish)
data("Social_Evolution")
callNetwork <- make_network(nodes = actors, directed = TRUE) |>
  link_events(change_event = calls, nodes = actors)
callsDependent <- define_dependent_events(
  events = calls, nodes = actors, default_network = call_network
)
socialEvolutionData <- make_data(callsDependent)
data_to_stan <- make_data_hmm(
  rate_effects = callsDependent ~ indeg + outdeg,
  choice_effects = NULL,
  data = socialEvolutionData
)

postProcess <- hmm_post_processing(data_to_stan, cmdstan_samples)
}
}
