% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{HMMDraws2LS}
\alias{HMMDraws2LS}
\title{Get Draws from the Posterior Distribution of a HMM-DyNAM}
\usage{
HMMDraws2LS(
  data2Stan,
  cmdstanSamples,
  kStates = data2Stan$dataStan$kS,
  type = c("2label.switching", "draws_matrix", "draws_array", "draws_df"),
  rescale = FALSE
)
}
\arguments{
\item{data2Stan}{a object of class \code{"goldfish.latent.data"} outcome of a call
to \code{\link[=CreateDataHMM]{CreateDataHMM()}}.}

\item{cmdstanSamples}{a object of class \code{"CmdStanMCMC"} with the posterior
samples of a HMM from the \code{data2Stan}.}

\item{kStates}{an integer with the number of states used to get samples
from the posterior distribution in \code{cmdstanSamples}.}

\item{type}{a character specifying weather the draws are shaped for the
\code{label.switching} package or a simple matrix format.}
}
\value{
a list with draws and additional information
}
\description{
Get Draws from the Posterior Distribution of a HMM-DyNAM
}
\examples{
\donttest{
library(goldfish)
data("Social_Evolution")
callNetwork <- defineNetwork(nodes = actors, directed = TRUE) |>
  linkEvents(changeEvent = calls, nodes = actors)
callsDependent <- defineDependentEvents(
  events = calls, nodes = actors, defaultNetwork = callNetwork
)
data2stan <- CreateDataHMM(
  rateEffects = callsDependent ~ indeg + outdeg,
  choiceEffects = NULL
)

drawsMCMC <- HMMDraws2LS(data2stan, cmdstanSamples)
}
}
