% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-data.R
\name{make_cox_model_data}
\alias{make_cox_model_data}
\title{Create a data frame and apply support constraint}
\usage{
make_cox_model_data(data, sub_model = c("rate", "choice"))
}
\arguments{
\item{data}{a list with stan data.}

\item{sub_model}{a character string indicating the sub_model,
either "rate" or "choice".}
}
\value{
a data frame with the data in long format.
}
\description{
Create a data frame and apply support constraint
}
