% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{make_data_hmm}
\alias{make_data_hmm}
\title{Create data for Stan Hidden Markov Model model}
\usage{
make_data_hmm(
  rate_effects,
  choice_effects,
  model = c("DyNAM"),
  k_states = 3,
  support_constraint = NULL,
  control_preprocessing = NULL,
  scale = TRUE,
  progress = getOption("progress"),
  data = NULL
)
}
\arguments{
\item{rate_effects}{a \code{formula} specification as in \code{\link[goldfish:estimate]{goldfish::estimate()}}.}

\item{choice_effects}{a \code{formula} specification as in \code{\link[goldfish:estimate]{goldfish::estimate()}}.}

\item{model}{Current version only support \code{"DyNAM"} model.}

\item{support_constraint}{a \code{formula} with only an effect that gives the
information of the restricted set to consider.
In the case of the \code{"choice"} sub-model, it corresponds to the choice set
available to received an event at each moment of time.
In the case of the \code{"rate"} sub-model, it corresponds to the competing set
available to send an event at each moment of time.
In the case of the \code{"REM"} model, it corresponds to the set of dyads
available to create an event at each moment of time.
The left hand side of the formula is left empty.
The effect should be coded in such a way that value 1 indicated
the availability and 0 the absence.
In the case that some actors left or join the process at any point of time is
better to use the \code{present} variable in the node data frame linking to it the
time varying changes of composition of the actors set.}

\item{control_preprocessing}{a \code{preprocessing_opt.goldfish} object
output from a call to \code{\link[goldfish:set_preprocessing_opt]{goldfish::set_preprocessing_opt()}} to compute
the update statistics of the event sequence.}

\item{scale}{logical value. Whether to standardize the effect stats at
the end.}

\item{progress}{logical argument passed to
\code{\link[goldfish:gather_model_data]{goldfish::gather_model_data()}} to show the progress of the preprocessing
of the events sequence.}

\item{data}{a \code{data.goldfish} object. Output of \code{\link[goldfish:make_data]{goldfish::make_data()}}.}
}
\value{
a list with the following components.
\describe{
\item{data_stan}{a list with the information necessary to run a HMC using
Stan.}
\item{names_effects}{a character vector with terms in the random and fixed
effects formulas and their final name.}
\item{effect_description}{an array with detailed and comprehensible
information of the terms used in the random and fixed effects formulas.}
\item{scale_stats}{a list with the statistics used for the standardization
when \code{scale = TRUE}.}
}
}
\description{
The first step is create the data in the structure expected by the \code{Stan}
code designed for the HMM-DyNAM.
Additional information of the effects used during preprocessing is added to
the return object.
}
\examples{
\donttest{
library(goldfish)
data("Social_Evolution")
callNetwork <- make_network(nodes = actors, directed = TRUE) |>
  link_events(change_event = calls, nodes = actors)
callsDependent <- define_dependent_events(
  events = calls, nodes = actors, default_network = callNetwork
)
socialEvolutionData <- make_data(callsDependent)
data2stan <- make_data_hmm(
  rate_effects = callsDependent ~ indeg + outdeg,
  choice_effects = callsDependent ~ recip + trans,
  data = socialEvolutionData
)
}
}
