% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_model_code}
\alias{make_model_code}
\title{Create a Stan code file}
\usage{
make_model_code(data_stan, ...)
}
\arguments{
\item{data_stan}{a \code{list} output of a \code{\link[=CreateData]{CreateData()}} call.}

\item{...}{additional arguments to be passed to
\code{\link[cmdstanr:write_stan_file]{cmdstanr::write_stan_file()}}}
}
\value{
The path to a file with \code{stan} extension.
It contains the code with the specification of data structure, priors,
and log-likelihood of the given model.
}
\description{
\code{goldfish.latent} offers working version of the models to work with Stan.
Users should have installed \code{cmdstanr} package and \code{CmdStan} before.
Follow the instructions instructions from \code{cmdstanr} documentation.
}
\details{
\code{cmdstanr} functionalities makes possible for users to write a local copy
of the Stan code and modify it for other purposes.
The default in \code{\link[cmdstanr:write_stan_file]{cmdstanr::write_stan_file()}} is to write the model in a
temporal folder. Using the \code{dir} argument is possible to write the model
in a folder specifies by the user.
}
\examples{
\donttest{
library(goldfish)
library(cmdstanr)
data("Social_Evolution")
callNetwork <- make_network(nodes = actors, directed = TRUE) |>
  link_events(change_event = calls, nodes = actors)
callsDependent <- define_dependent_events(
  events = calls, nodes = actors, default_network = callNetwork
)
socialEvolutionData <- make_data(callsDependent)
data2stan <- make_data_re(
  random_effects = list(inertia ~ 1),
  fixed_effects = callsDependent ~ recip + trans,
  data = data
)

stanCode <- make_model_code(data2stan)
}
}
