% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{permute_mcmc_array}
\alias{permute_mcmc_array}
\title{Permute MCMC Array}
\usage{
permute_mcmc_array(
  draws,
  permutation,
  k_states,
  model = c("DNHMM", "DNCHMM", "DNCHMMRE"),
  sub_model = c("choice", "rate", "both"),
  hmm_type = c("discrete", "continuous")
)
}
\arguments{
\item{draws}{a list output of \code{HMMDraws2LS}}

\item{permutation}{a numeric vector output from the \code{label.switching}
package with the permutation to apply to each draw}

\item{kStates}{the number of states in the HMM}

\item{subModel}{a character specifying the submodel of the HMM}

\item{hmmType}{a character specifying the type of HMM between
discrete and continuous}
}
\value{
a data frame with the permuted draws
}
\description{
Permute MCMC Array
}
