% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_data}
\alias{sample_data}
\title{sample preprocessed data}
\usage{
sample_data(
  data,
  fraction_set = 1,
  method_set = c("srswor", "systematic"),
  fraction_events = 1,
  method_events = c("systematic", "srswor")
)
}
\arguments{
\item{data}{a \code{goldfish.latent.data} object.}

\item{fraction_set}{a numerical value greater than 0. If \code{fraction_set} is
between 0 and 1, it is interpreted as the proportion of the choice set
to sample. If \code{fraction_set} is greater than 1, it is interpreted as the
number of alternatives to sample.}

\item{method_set}{a character string indicating the sampling method for the
choice set.}

\item{fraction_events}{a numerical value greater than 0. If
\code{fraction_events} is between 0 and 1, it is interpreted as the proportion
of events to sample. If \code{fraction_events} is greater than 1, it is
interpreted as the number of events to sample.}

\item{method_events}{a character string indicating the sampling method for
the events.}
}
\value{
a \code{goldfish.latent.data} object with the sampled data.
}
\description{
For arguments \code{methodChoiceSet} and \code{methodEvents} the available methods are
Sample preprocessed data
}
\details{
For arguments \code{method_set} and \code{method_events} the available methods are
\code{"systematic"} and \code{"srswor"}. They correspond to
systematic sampling and simple random sampling without replacement,
respectively. It is possible to use an external function that has arguments
\code{N} and \code{fraction} and return a numerical vector of the samples
to keep.
}
\examples{
# sampledData <- sample_data(data)
}
