% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{transformMCMCArray}
\alias{transformMCMCArray}
\title{Transform MCMC Array}
\usage{
transformMCMCArray(
  data2Stan,
  cmdstanSamples,
  labelSwitching,
  postProcessing,
  kStates = data2Stan$dataStan$kS,
  method,
  rescale = TRUE
)
}
\arguments{
\item{data2Stan}{a list output from \code{make_data_hmm}}

\item{cmdstanSamples}{a list output from \code{CmdStanMCMC}}

\item{labelSwitching}{a list output from \code{label.switching}}

\item{postProcessing}{a list output from \code{hmm_post_processing}}

\item{kStates}{the number of states in the HMM}

\item{method}{a character specifying the method to use for the permutation}

\item{rescale}{a logical specifying if the draws should be rescaled}
}
\value{
a list with the permuted post processed statistics
}
\description{
Transform MCMC Array
}
