% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffects-actor.R
\name{compute_log_likelihood}
\alias{compute_log_likelihood}
\title{Compute log-likelihood using MCMC samples}
\usage{
compute_log_likelihood(
  cmdstan_samples,
  data_stan,
  type = c("marginal", "conditional"),
  n_nodes = ifelse(type == "marginal", 11L, NULL),
  split_size = NULL,
  spec = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{cmdstan_samples}{a \code{draws_array} or a \code{CmdStanFit} object with MCMC
samples from the posterior distribution. In the case of a \code{draws_array}
object is expected to have three dimensions corresponding to iterations,
chain and variables.}

\item{data_stan}{a \code{list} output of a \code{\link[=CreateData]{CreateData()}} call.}

\item{type}{a \code{character} value. It indicates whether the log-likelihood
computation should return the \code{"conditional"} or the \code{"marginal"} version.}

\item{n_nodes}{an \code{integer}. The number of quadrature point to use in the
Gauss-Hermite quadrature approximation use to integrate out the
random-effects.}

\item{split_size}{an \code{integer} or \code{NULL}. It is use when the \code{type} is
\code{"conditional"}. When it is \code{NULL}, the \code{split_size} is set to have
roughly \code{4e4} rows sent to a processor when
the matrix \code{X}, containing the change statistics, has more
than \code{1e6}rows. If \code{X} has less than \code{1e6} rows, the \code{split_size} is set
in such way that every processor would have the same amount of rows to
process.
The default value is \code{NULL}.}

\item{spec}{A specification appropriate to the type of cluster, see
\code{\link[parallel:makeCluster]{parallel::makeCluster()}} for a detail description. In the simplest case,
an integer defining the number of processors to use during the parallel
computation.}

\item{...}{Additional arguments and options to be passed to the
\code{\link[parallel:makeCluster]{parallel::makeCluster()}} call.}
}
\value{
An array with the log-likelihood for each event when
\code{type = "conditional"} or the log-likelihood for each sender actor after
the random effects are integrated out when \code{type = "marginal"}.
}
\description{
The function computes the log-likelihood for each MCMC sample from the
posterior distribution.
}
\details{
Argument \code{type} allows to compute the conditional or marginal version of the
log-likelihood. The marginal version uses a Gauss-Hermite quadrature
approximation to integrate out the random effects
\insertCite{Merkle2019}{goldfish.latent}. The code is an adaptation from
the supplementary material of
\insertCite{Merkle2019;textual}{goldfish.latent}.
}
\examples{
\donttest{
library(goldfish)
library(cmdstanr)
data("Social_Evolution")
callNetwork <- make_network(nodes = actors, directed = TRUE) |>
  link_events(change_event = calls, nodes = actors)
callsDependent <- define_dependent_events(
  events = calls, nodes = actors, default_network = callNetwork
)
socialEvolutionData <- make_data(callsDependent)
data2stan <- make_data_re(
  random_effects = list(inertia ~ 1),
  fixed_effects = callsDependent ~ recip + trans,
  data = socialEvolutionData
)

stanCode <- make_model_code(data2stan)

mod01 <- cmdstan_model(stanCode)
mod01Samples <- mod01$sample(
  data = data2stan[["data_stan"]],
  parallel_chains = 4, chains = 4,  iter_warmup = 500, iter_sampling = 500,
  show_messages = FALSE
)

margLogLikMod01 <- compute_log_likelihood(mod01Samples, data2stan, spec = 4)
condLogLikMod01 <- compute_log_likelihood(mod01Samples, data2stan,
                                        type = "conditional", spec = 4)
}
}
\references{
\insertRef{Merkle2019}{goldfish.latent}
}
