% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{CreateDataHMM}
\alias{CreateDataHMM}
\title{Create data for Stan Hidden Markov Model model}
\usage{
CreateDataHMM(
  rateEffects,
  choiceEffects,
  model = c("DyNAM", "REM"),
  kRegimes = 3,
  supportConstraint = NULL,
  preprocessArgs = NULL,
  scale = FALSE,
  progress = getOption("progress"),
  envir = new.env()
)
}
\arguments{
\item{rateEffects}{a \code{formula} specification as in \code{\link[goldfish:estimate]{goldfish::estimate()}}.}

\item{choiceEffects}{a \code{formula} specification as in \code{\link[goldfish:estimate]{goldfish::estimate()}}.}

\item{model}{Current version only support \code{"DyNAM"} model, enhancements
on the code would allow to use \code{"REM"} model too.}

\item{supportConstraint}{a \code{formula} with only an effect that gives the
information of the alters available to received an event at each moment
of time. The left hand side of the formula is left empty.
It usually will correspond to a call of the \code{tie(network)} where \code{network}
is a binary adjacency matrix where the value 1 indicated that actor $j$ is
available to receive an event from actor $i$.
In the case that some actors left or join the process at any point of time is
better to use the \code{present} variable in the node data frame linking to it the
time varying changes of composition of the actors set.}

\item{preprocessArgs}{a list with additional preprocess arguments used by
\code{\link[goldfish:GatherPreprocessing]{goldfish::GatherPreprocessing()}} to compute the changes statistics of the
event sequence.}

\item{scale}{logical value. Whether to standardize the effect stats at
the end.}

\item{progress}{logical argument passed to \code{\link[goldfish:GatherPreprocessing]{goldfish::GatherPreprocessing()}}
to show the progress of the preprocessing of the events sequence.}

\item{envir}{an \code{environment} where the formula objects are available.}
}
\value{
a list with the following components.
\describe{
\item{dataStan}{a list with the information necessary to run a HMC using
Stan.}
\item{namesEffects}{a character vector with terms in the random and fixed
effects formulas and their final name.}
\item{effectDescription}{an array with detailed and comprehensible
information of the terms used in the random and fixed effects formulas.}
\item{scale}{a list with the statistics used for the standardization when
\code{scale = TRUE}.}
}
}
\description{
The first step is create the data in the structure expected by the \code{Stan}
code designed for the HMM-DyNAM.
Additional information of the effects used during preprocessing is added to
the return object.
}
\examples{
\donttest{
library(goldfish)
data("Social_Evolution")
callNetwork <- defineNetwork(nodes = actors, directed = TRUE) |>
  linkEvents(changeEvent = calls, nodes = actors)
callsDependent <- defineDependentEvents(
  events = calls, nodes = actors, defaultNetwork = callNetwork
)
data2stan <- CreateDataHMM(
  rateEffects = callsDependent ~ indeg + outdeg,
  choiceEffects = callsDependent ~ recip + trans
)
}
}
