% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldfish.latent-package.R
\docType{package}
\name{goldfish.latent-package}
\alias{goldfish.latent-package}
\alias{goldfish.latent}
\title{goldfish.latent: Relational Event Models with Latent Variables}
\description{
Bayesian estimation of relational event models with latent variables. The class of models introduced in the package extends the dynamic network actor models ('DyNAMs') and relational event models ('REMs') estimated with the package goldfish. Bayesian inference methods as hamiltonian markov chain (HMC) as implemented in Stan are used to perform inferences of the proposed models.

Bayesian estimation of relational event models with latent variables.
The class of models introduced in the package extends the
dynamic network actor models ('DyNAMs') and relational event models ('REMs')
estimated with the package goldfish.
Bayesian inference methods as hamiltonian markov chain (HMC) as implemented
in Stan are used to perform inferences of the proposed models.
}
\references{
Uzaheta A., Amati V. and Stadtfeld, C. (2023). Random effects in
dynamic network actor models.
\emph{Network Science}. \doi{10.1017/nws.2022.37}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/auzaheta/goldfish.latent}
  \item Report bugs at \url{https://github.com/auzaheta/goldfish.latent/issues}
}


\code{\link[goldfish:estimate]{goldfish::estimate()}}
}
\author{
\strong{Maintainer}: Alvaro Uzaheta \email{ualvaro@ethz.ch} (\href{https://orcid.org/0000-0003-4367-3670}{ORCID})

}
\keyword{internal}
