% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_data}
\alias{sample_data}
\title{Sample preprocessed data}
\usage{
sample_data(
  data,
  fraction_set = 1,
  sample_size_set = NULL,
  method_set = c("srswor", "systematic"),
  fraction_events = 1,
  sample_size_events = NULL,
  method_events = c("systematic", "srswor")
)
}
\arguments{
\item{data}{a \code{goldfish.latent.data} object.}

\item{fraction_set}{a numerical value between 0 and 1.
It is the proportion of cases to sample. If \code{sample_size_set} is not
\code{NULL}, it is ignored.}

\item{sample_size_set}{an integerish value greater than 0.
It defines the number of cases to sample. Ignores \code{fraction_set} if
\code{sample_size_set} is not \code{NULL}.}

\item{method_set}{a character string indicating the sampling method for the
choice set.}

\item{fraction_events}{a numerical value between 0 and 1.
It is the proportion of events to sample. If \code{sample_size_events} is not
\code{NULL}, it is ignored.}

\item{sample_size_events}{an integerish value greater than 0.
It defines the number of events to sample. Ignores \code{fraction_events} if
\code{sample_size_events} is not \code{NULL}.}

\item{method_events}{a character string indicating the sampling method for
the events.}
}
\value{
a \code{goldfish.latent.data} object with the sampled data and the
selection probabilities of the events/cases sampled.
}
\description{
For arguments \code{method_set} and \code{method_events} the available methods are
\code{"systematic"} and \code{"srswor"}. They correspond to
systematic sampling and simple random sampling without replacement,
respectively. It is possible to use an external function that has arguments
\code{N} and \code{fraction} and return a numerical vector of the samples
to keep.
}
\examples{
sampledData <- sample_data(data)
}
