% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_data}
\alias{sample_data}
\title{sample preprocessed data}
\usage{
sample_data(
  data,
  fraction_choice_set = 0.1,
  method_choice_set = c("srswor", "systematic"),
  fraction_events = 1,
  method_events = c("systematic", "srswor")
)
}
\arguments{
\item{data}{output from \code{\link[=make_data_hmm]{make_data_hmm()}} or \code{\link[=make_data_re]{make_data_re()}}}

\item{fraction_choice_set}{numerical value that indicates the proportion of
alternatives to sample from the choice set or the compiting actors in the
choice and rate model, respectively.}

\item{method_choice_set}{character value indicating the function name used to
generate the sample.}

\item{fraction_events}{numerical value that indicates the proportion of
events to sample. When the \code{data} object contains data for both sub-models,
a pair sample is selected, i.e., the sample events contains information of
the rate and choice sub-models.}

\item{method_events}{character value indicating the function name used to
generate the sample.}
}
\value{
an object with the same meta information as \code{data} with the sampled
events and alternative choices.
}
\description{
For arguments \code{methodChoiceSet} and \code{methodEvents} the available methods are
\code{"systematic"} and \code{"srswor"}. They correspond to
systematic sampling and simple random sampling without replacement,
respectively. It is possible to use an external function that has arguments
\code{N} and \code{fraction} and return a numerical vector of the samples
to keep.
}
\examples{
sampledData <- sample_data(data)
}
