% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-data.R
\name{write_json}
\alias{write_json}
\title{Write model data to a JSON file}
\usage{
write_json(x, file_name, n_chunks = 10)
}
\arguments{
\item{x}{Data object}

\item{file_name}{File name to write the JSON string to,
needs the .json extension.}

\item{n_chunks}{Number of chunks to divide the JSON string to avoid
the error.}
}
\description{
When the data is too large, the JSON string can be too long and cause
and error when writing the data to a file
(R character strings are limited to 2^31-1 bytes).
In that case, the JSON string is divided into chunks.
}
